fix: Use pulls API instead of Github Search (#81)
I ran into the secondary rate limit issue in this Github Actions. From
deeper investigation, it appears the code for fetching PRs has a tighter
rate limit than the past.

From investigation, it appears that switching to the pulls API should
avoid the rate limit on the search API.

This should fix #73.